requirements.txt
# Backend dependencies
fastapi==0.111.0
uvicorn[standard]==0.23.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
google-cloud-storage==2.10.0

# ML dependencies
pandas==2.1.0
numpy
scikit-learn==1.6.1
xgboost
joblib==1.3.2
pydantic
